+++ app/src/main/java/com/fibelatti/photowidget/configure/BackgroundRestrictedSheetDialog.kt
@@ -1,5 +1,8 @@
 package com.fibelatti.photowidget.configure
 
+import android.content.Context
+import android.os.PowerManager
+import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
@@ -19,6 +22,7 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.material3.TextButton
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
@@ -27,9 +31,10 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.DialogProperties
+import androidx.core.content.getSystemService
 import com.fibelatti.photowidget.R
 import com.fibelatti.photowidget.platform.appSettingsIntent
-import com.fibelatti.photowidget.platform.batteryUsageSettingsIntent
+import com.fibelatti.photowidget.platform.disableBatteryOptimizationIntent
 import com.fibelatti.ui.foundation.AppBottomSheet
 import com.fibelatti.ui.foundation.AppSheetState
 import com.fibelatti.ui.foundation.TextWithLinks
@@ -98,17 +103,21 @@ fun BackgroundRestrictionWarningDialog(
 fun BackgroundRestrictionBottomSheet(
     sheetState: AppSheetState,
 ) {
-    val localContext = LocalContext.current
+    val localContext: Context = LocalContext.current
+    val powerManager: PowerManager? = remember { localContext.getSystemService<PowerManager>() }
+    val isBatteryUsageRestricted: Boolean = powerManager
+        ?.isIgnoringBatteryOptimizations(localContext.packageName) != true
 
     AppBottomSheet(
         sheetState = sheetState,
     ) {
         BackgroundPickerContent(
+            isBatteryUsageRestricted = isBatteryUsageRestricted,
+            onGrantPermissionClick = {
+                localContext.startActivity(disableBatteryOptimizationIntent(context = localContext))
+            },
             onOpenAppSettingsClick = {
                 localContext.startActivity(appSettingsIntent(context = localContext))
-            },
-            onOpenPowerOptimizationSettings = {
-                localContext.startActivity(batteryUsageSettingsIntent())
             },
         )
     }
@@ -117,8 +126,9 @@ fun BackgroundRestrictionBottomSheet(
 @Composable
 @OptIn(ExperimentalMaterial3ExpressiveApi::class)
 private fun BackgroundPickerContent(
+    isBatteryUsageRestricted: Boolean,
+    onGrantPermissionClick: () -> Unit,
     onOpenAppSettingsClick: () -> Unit,
-    onOpenPowerOptimizationSettings: () -> Unit,
 ) {
     Column(
         modifier = Modifier
@@ -158,20 +168,25 @@ private fun BackgroundPickerContent(
 
         Spacer(modifier = Modifier.size(8.dp))
 
+        AnimatedVisibility(
+            visible = isBatteryUsageRestricted,
+            modifier = Modifier.fillMaxWidth(),
+        ) {
+            FilledTonalButton(
+                onClick = onGrantPermissionClick,
+                shapes = ButtonDefaults.shapes(),
+                modifier = Modifier.fillMaxWidth(),
+            ) {
+                Text(text = stringResource(R.string.restriction_warning_grant_battery_permission))
+            }
+        }
+
         FilledTonalButton(
             onClick = onOpenAppSettingsClick,
             shapes = ButtonDefaults.shapes(),
             modifier = Modifier.fillMaxWidth(),
         ) {
             Text(text = stringResource(R.string.restriction_warning_open_app_settings))
-        }
-
-        FilledTonalButton(
-            onClick = onOpenPowerOptimizationSettings,
-            shapes = ButtonDefaults.shapes(),
-            modifier = Modifier.fillMaxWidth(),
-        ) {
-            Text(text = stringResource(R.string.restriction_warning_open_power_settings))
         }
     }
 }
@@ -181,8 +196,9 @@ private fun BackgroundPickerContent(
 private fun BackgroundPickerContentPreviews() {
     ExtendedTheme {
         BackgroundPickerContent(
+            isBatteryUsageRestricted = true,
+            onGrantPermissionClick = {},
             onOpenAppSettingsClick = {},
-            onOpenPowerOptimizationSettings = {},
         )
     }
 }